Modifications according to the suggestions
